+++ src/lib/authentication-error.ts
